+++ src/format/codec_config.rs
@@ -575,7 +575,7 @@ impl CodecConfig {
             .map(|p| {
                 p.spec
                     .format
-                    .tele_event_max
+                    .telephone_event_max
                     .unwrap_or(FormatParams::DEFAULT_TELE_EVENT_MAX)
             })
     }